+++ Items/TubularMagiluminescence/TubularMagiluminescence.cs
@@ -12,7 +12,7 @@ namespace Combinations.Items.TubularMagiluminescence
         public override void SetStaticDefaults()
         {
             Tooltip.SetDefault("Grants the ability to float in water\n" +
-                "Increases movement speed an acceleration\n" +
+                "Increases movement speed and acceleration\n" +
                 "Provides light when worn\n" +
                 "'How exactly does this float?'");
         }